refactor auth, services
