+++ frontend/lib/main.dart
@@ -7,8 +7,8 @@ import 'package:video_player_media_kit/video_player_media_kit.dart';
 import 'package:window_manager/window_manager.dart';
 import 'package:inter_sign/themes/light_theme.dart';
 
-import 'auth/auth_gate.dart';
-import 'auth/auth_service.dart';
+import 'screens/auth_gate.dart';
+import 'services/auth_service.dart';
 
 void main() async {
   WidgetsFlutterBinding.ensureInitialized();
