+++ frontend/lib/screens/auth/login_screen.dart
@@ -4,7 +4,7 @@ import 'package:inter_sign/screens/primary/dashboard.dart';
 import 'package:inter_sign/widgets/info_card.dart';
 import 'package:inter_sign/widgets/logo_widget.dart';
 
-import '../../auth/auth_service.dart';
+import '../../services/auth_service.dart';
 import '../../utils/responsive.dart';
 import '../../utils/show_toast.dart';
 import '../../widgets/form_container.dart';
