+++ frontend/lib/screens/auth_gate.dart
@@ -1,8 +1,7 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_secure_storage/flutter_secure_storage.dart';
 import 'package:inter_sign/screens/auth/login_screen.dart';
 import 'package:inter_sign/screens/primary/dashboard.dart';
-import 'package:inter_sign/auth/auth_service.dart';
+import 'package:inter_sign/services/auth_service.dart';
 import 'package:provider/provider.dart';
 
 class AuthGate extends StatelessWidget {
